requirements.txt
numpy<2.0
scipy
xarray
vtk
PySide6-essentials
pyqtgraph
requests
netCDF4
fpdf2
pdfrw
Pillow
PyYAML  #yaml
matplotlib
rapidfuzz
PySide6-QtAds
wavespectra
waveresponse
mafredo
nooverlap
